Move out input validation into base image processor
