+++ src/transformers/models/glpn/image_processing_glpn.py
@@ -19,19 +19,15 @@ from typing import List, Optional, Union
 import numpy as np
 import PIL.Image
 
-from ...image_processing_utils import BaseImageProcessor, BatchFeature
+from ...image_processing_utils import BaseImageProcessor, BatchFeature, validate_preprocess_arguments
 from ...image_transforms import resize, to_channel_dimension_format
 from ...image_utils import (
     ChannelDimension,
     PILImageResampling,
     get_image_size,
     infer_channel_dimension_format,
-    is_scaled_image,
     make_list_of_images,
     to_numpy_array,
-    valid_images,
-    validate_kwargs,
-    validate_preprocess_arguments,
 )
 from ...utils import TensorType, logging
 
@@ -82,6 +78,15 @@ class GLPNImageProcessor(BaseImageProcessor):
             "data_format",
             "input_data_format",
         ]
+
+    def _validate_preprocess_arguments(self, do_resize, size_divisor, resample):
+        # rescale() uses a constant rescale_factor. It does not need to be validated
+        # with a rescale_factor.
+        validate_preprocess_arguments(
+            do_resize=do_resize,
+            size=size_divisor,  # size_divisor is used as a parameter for optimal resizing instead of size.
+            resample=resample,
+        )
 
     def resize(
         self,
@@ -187,30 +192,8 @@ class GLPNImageProcessor(BaseImageProcessor):
 
         images = make_list_of_images(images)
 
-        validate_kwargs(captured_kwargs=kwargs.keys(), valid_processor_keys=self._valid_processor_keys)
-
-        if not valid_images(images):
-            raise ValueError(
-                "Invalid image type. Must be of type PIL.Image.Image, numpy.ndarray, "
-                "torch.Tensor, tf.Tensor or jax.ndarray."
-            )
-
-        # Here, the rescale() method uses a constant rescale_factor. It does not need to be validated
-        # with a rescale_factor.
-        validate_preprocess_arguments(
-            do_resize=do_resize,
-            size=size_divisor,  # Here, size_divisor is used as a parameter for optimal resizing instead of size.
-            resample=resample,
-        )
-
         # All transformations expect numpy arrays.
         images = [to_numpy_array(img) for img in images]
-
-        if is_scaled_image(images[0]) and do_rescale:
-            logger.warning_once(
-                "It looks like you are trying to rescale already rescaled images. If the input"
-                " images have pixel values between 0 and 1, set `do_rescale=False` to avoid rescaling them again."
-            )
 
         if input_data_format is None:
             # We assume that all images have the same channel dimension format.
