+++ src/transformers/models/seggpt/image_processing_seggpt.py
@@ -27,10 +27,8 @@ from ...image_utils import (
     ImageInput,
     PILImageResampling,
     infer_channel_dimension_format,
-    is_scaled_image,
     make_list_of_images,
     to_numpy_array,
-    valid_images,
 )
 from ...utils import TensorType, is_torch_available, is_vision_available, logging, requires_backends
 
@@ -328,21 +326,6 @@ class SegGptImageProcessor(BaseImageProcessor):
         # If segmentation map is passed we expect 2D images
         images = make_list_of_images(images, expected_ndims=2 if do_convert_rgb else 3)
 
-        if not valid_images(images):
-            raise ValueError(
-                "Invalid image type. Must be of type PIL.Image.Image, numpy.ndarray, "
-                "torch.Tensor, tf.Tensor or jax.ndarray."
-            )
-
-        if do_resize and size is None:
-            raise ValueError("Size must be specified if do_resize is True.")
-
-        if do_rescale and rescale_factor is None:
-            raise ValueError("Rescale factor must be specified if do_rescale is True.")
-
-        if do_normalize and (image_mean is None or image_std is None):
-            raise ValueError("Image mean and std must be specified if do_normalize is True.")
-
         # All transformations expect numpy arrays.
         images = [to_numpy_array(image) for image in images]
 
@@ -352,7 +335,7 @@ class SegGptImageProcessor(BaseImageProcessor):
                 " images have pixel values between 0 and 1, set `do_rescale=False` to avoid rescaling them again."
             )
 
-        if input_data_format is None and not do_convert_rgb:
+        if input_data_format is None and not is_mask:
             # We assume that all images have the same channel dimension format.
             input_data_format = infer_channel_dimension_format(images[0])
